Recommiting with a proper commit message
I added basic methods to weapon, player, powerup, projectile.
Created GamePanel class that draws our field and modified the GUI class to use it instead. So far the GamePanel does what it does before(draw a black background) plus it draws a player rectangle and implements a KeyListener to move it around